redditmatcher/requirements.txt
streamlit
python-dotenv
requests
praw
google-generativeai
scikit-learn
matplotlib
nltk
bertopic
hdbscan